ClientAPI/requirements.txt
brotli==1.0.9
brotlipy==0.7.0
ConfigParser==5.3.0
cryptography==38.0.0
Cython==0.29.23
dates==0.0.1
dl==0.1.0
docutils==0.17
Flask==2.1.3
HTMLParser==0.0.2
importlib_metadata==4.12.0
ipaddr==2.2.0
keyring==22.3.0
lockfile==0.12.2
lxml==4.6.3
mypy_extensions==0.4.3
ordereddict==1.1
polygon==1.1.0
protobuf==4.21.9
pycryptodome==3.15.0
pyOpenSSL==22.1.0
rsa==4.9
typing_extensions==3.7.4.3
zipp==3.8.1
gunicorn==20.1.0
python-dotenv